refactor: 디테일 페이지 구조 개선
- 디테일 페이지의 구조를 개선합니다.
