+++ Components/Detail/DetailContent.tsx
@@ -10,7 +10,6 @@ interface DetailContentProps {
   content: string;
 }
 
-// 다이나믹 임포트 고려
 const DetailContent = ({ content }: DetailContentProps) => {
   return (
     <Viewer
