+++ package.json
@@ -29,6 +29,7 @@
     "react": "18.2.0",
     "react-datepicker": "^4.10.0",
     "react-dom": "18.2.0",
+    "react-spinners": "^0.13.8",
     "react-supabase": "^0.2.0",
     "recoil": "^0.7.6",
     "styled-components": "^5.3.6",
@@ -36,9 +37,9 @@
     "typescript": "4.9.5"
   },
   "devDependencies": {
-    "@types/prismjs": "^1.26.0",
     "@types/lodash": "^4.14.191",
     "@types/node": "18.13.0",
+    "@types/prismjs": "^1.26.0",
     "@types/react": "18.0.27",
     "@types/react-datepicker": "^4.8.0",
     "@types/react-dom": "18.0.10",
