+++ utils/APIs/supabase/getAllPosts.ts
@@ -1,5 +1,8 @@
 import supabase from "@/lib/supabase";
 
+// TODO: Add pagination
+// TODO: 필요한 데이터만 가져오도록 수정
+
 const getAllPosts = async () => {
   const res = await supabase
     .from("post")
